fix(file): correct html path generation
Restructure HTML saving to fix path bugs where files were saved as
directories and improve workspace functionality with proper nested path
handling.
