+++ src/file/mod.rs
@@ -3,5 +3,7 @@ mod ops;
 mod workspace;
 
 pub use error::{FileError, PathError};
-pub use ops::{load_image, load_json, load_toml, make_directory, save_json, save_page};
+pub use ops::{
+    load_image, load_json, load_toml, make_directory, save_html, save_json, url_to_file_name,
+};
 pub use workspace::Workspace;
